# BookLibraryApi/bin/Release/netcoreapp3.1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BookLibraryApi",
    "version": "1.0"
  },
  "paths": {
    "/api/authors": {
      "get": {
        "tags": [
          "Author"
        ],
        "operationId": "GetAuthors",
        "parameters": [
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Author"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Author"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AuthorCreation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Author"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/authors/{id}": {
      "get": {
        "tags": [
          "Author"
        ],
        "operationId": "GetAuthor",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Author"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Author"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateAuthor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Author"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateAuthorJsonPatchDocument"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Author"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Genres/{GenreId}/Books": {
      "get": {
        "tags": [
          "Book"
        ],
        "operationId": "GetBooks",
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Author",
            "type": "string"
          },
          {
            "in": "query",
            "name": "YearOfPublish",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Author",
            "type": "string"
          },
          {
            "in": "query",
            "name": "YearOfPublish",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookCreation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Genres/{GenreId}/Books/{BookId}": {
      "get": {
        "tags": [
          "Book"
        ],
        "operationId": "GetBook",
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "BookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "BookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "fields",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Book"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "BookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Book"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "BookId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookUpdateJsonPatchDocument"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "BookId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AuthorCreation": {
      "required": [
        "bio",
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "pictureUrl": {
          "format": "uri",
          "type": "string"
        },
        "bio": {
          "maxLength": 1000,
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date-time",
          "type": "string"
        },
        "dateOfDeath": {
          "format": "date-time",
          "type": "string"
        },
        "genreId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "BookCreation": {
      "required": [
        "bookTitle",
        "description"
      ],
      "type": "object",
      "properties": {
        "bookTitle": {
          "type": "string"
        },
        "bookCover": {
          "format": "uri",
          "type": "string"
        },
        "dateOfPublish": {
          "format": "date-time",
          "type": "string"
        },
        "numberOfBookPages": {
          "format": "int32",
          "type": "integer"
        },
        "publisher": {
          "type": "string"
        },
        "description": {
          "maxLength": 1000,
          "type": "string"
        },
        "authorId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "BookUpdate": {
      "required": [
        "bookTitle",
        "description"
      ],
      "type": "object",
      "properties": {
        "bookTitle": {
          "type": "string"
        },
        "bookCover": {
          "format": "uri",
          "type": "string"
        },
        "dateOfPublish": {
          "format": "date-time",
          "type": "string"
        },
        "numberOfBookPages": {
          "format": "int32",
          "type": "integer"
        },
        "publisher": {
          "type": "string"
        },
        "description": {
          "maxLength": 1000,
          "type": "string"
        },
        "authorId": {
          "format": "uuid",
          "type": "string"
        },
        "genreId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "BookUpdateJsonPatchDocument": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BookUpdateOperation"
          },
          "readOnly": true
        },
        "contractResolver": {
          "$ref": "#/definitions/IContractResolver"
        }
      }
    },
    "BookUpdateOperation": {
      "type": "object",
      "properties": {
        "operationType": {
          "$ref": "#/definitions/OperationType"
        },
        "path": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "value": { }
      }
    },
...
    "IContractResolver": {
      "type": "object"
    },
    "OperationType": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4,
        5,
        6
      ],
      "type": "integer"
    },
...
    "UpdateAuthor": {
      "required": [
        "bio",
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "pictureUrl": {
          "format": "uri",
          "type": "string"
        },
        "bio": {
          "maxLength": 1000,
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date-time",
          "type": "string"
        },
        "dateOfDeath": {
          "format": "date-time",
          "type": "string"
        },
        "genreId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "UpdateAuthorJsonPatchDocument": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UpdateAuthorOperation"
          },
          "readOnly": true
        },
        "contractResolver": {
          "$ref": "#/definitions/IContractResolver"
        }
      }
    },
    "UpdateAuthorOperation": {
      "type": "object",
      "properties": {
        "operationType": {
          "$ref": "#/definitions/OperationType"
        },
        "path": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "value": { }
      }
    }
  }
}